Split contents of index into multiple components
